backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
google-auth==2.29.0
python-dotenv==1.0.1
httpx==0.27.0
openai>=1.0.0
